Fixes for x86-64 in some classes
git-svn-id: https://svn.aros.org/svn/aros/trunk@38066 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/usb/classes/hub/hub.class.c
@@ -1144,7 +1144,7 @@ struct PsdDevice * GM_UNIQUENAME(nConfigurePort)(struct NepClassHub *nch, UWORD 
                 Permit();
                 /* Hub reference */
                 psdSetAttrs(PGA_DEVICE, pd,
-                            DA_HubDevice, (ULONG) nch->nch_Device,
+                            DA_HubDevice, nch->nch_Device,
                             DA_IsConnected, TRUE,
                             DA_AtHubPortNumber, port,
                             TAG_END);
